Add customed clang-tidy check for unused parameters
Added a custom clang-tidy check named UnusedParameterCustomCheck.

This implementation of an unused parameter check is inferior to the official
UnusedParameterCheck that already exists within clang-tidy. In fact, it is
very similar in implementation. This check was written for the exploration of
the Clang/LLVM compiler toolkit and to gather a greater understanding of how
clang-tidy traverses and manipulations the Abstract Syntax Tree (AST).

Signed-off-by: Anderson <achauph@sandia.gov>